requirements.txt

celery==5.4.0

Django==4.2.15
django-jalali-date==1.1.3
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1

flower==2.0.1
gunicorn==23.0.0

jdatetime==5.0.0
mysqlclient==2.2.4
openpyxl==3.1.5

persian-tools==0.0.11
phonenumbers==8.13.44
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
Unidecode==1.3.8
gunicorn==23.0.0
